Move code out of libratom/utils
